migrated tests, fixed snapshots, removed duplicate enzyme test file
